Allow the CucumberPro formatter to be enabled without a token.
This makes it easier to add the formatter without having to
conditionally enable it. If CUCUMBER_PRO_TOKEN is not defined,
the formatter does not check if the working copy is clean,
and it does not connect to the server. This is how the Cucumber-JVM
implementation works now.

When Cucumber-Ruby is able to read command line options from a
simple file (cucumber.opts), it will be a lot easier to automate
the installation of the formatter.

This fixes #2.